[fix] No more offset for text layers in SVG Template compared to PSD Template
I found a formula to calculate a Y Offset to apply to a layer based on its font size to take it back to the right place on Y.
For more versatility, one should apply this technic to the x pos, as the text layers are also offsetted on x (but it's less annoying) and to rotated text layers.
What I did was that I manually offseted back all the layers, wrote down the delta Y, gave all the Delta Y and Font Size to chat GPT, and asked it to find a linear relation between them. Someone who remembers middle school's level math could probably do it without Chat GPT (self burn)